feat: create modalDropdown and chage icon
