+++ src/components/header/header.interface.ts
@@ -2,10 +2,3 @@ export interface IHeader {
   switchWeekday: React.Dispatch<React.SetStateAction<number>>;
   weekDays: string[];
 }
-
-export interface IDropdown {
-  setNavSwitch: React.Dispatch<React.SetStateAction<boolean>>;
-  switchWeekday: React.Dispatch<React.SetStateAction<number>>;
-  dropdownItems: string[];
-  navSwitch: boolean;
-}
