+++ src/components/header/views/headerDropdown.interface.ts
@@ -1,0 +1,6 @@
+export interface IDropdown {
+  setNavSwitch: React.Dispatch<React.SetStateAction<boolean>>;
+  switchWeekday: React.Dispatch<React.SetStateAction<number>>;
+  dropdownItems: string[];
+  navSwitch: boolean;
+}
